requirements.txt
Django==4.2.7
django-ckeditor==6.7.0
Pillow==10.0.1